Revert "2022: functional programming?"
This reverts commit f3bc6f4ed7919781a3d58b6917ef693c61bbb252.

It turns out (thanks IRC!) that Fennel/Lua does not optimize away copies of
the tables here, so tbl.update (which will be in the hot path a lot) takes
linear time, which makes the frequency counting algorithm I did quadratic. Lua
(and hence Fennel) isn't really appropriate for this, so I'm going back to
an iterative style, at least for building up data structures.